requirements.txt
matplotlib==3.2.2
numpy==1.18.5
pandas==1.0.5
Pillow==7.2.0
scikit-learn==0.23.1
torch==1.6.0
torchvision==0.7.0
xgboost==1.2.0
